requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AWS
boto3>=1.34.0
botocore>=1.34.0

# Vector Database
qdrant-client>=1.7.0

# Document Processing
beautifulsoup4>=4.12.0
pandas>=2.1.0
unstructured>=0.11.0
pypdf>=3.17.0
python-multipart>=0.0.6

# LangChain and Text Processing
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
typing-extensions>=4.8.0

# Additional dependencies
pillow>=10.1.0
pytesseract>=0.3.10
numpy>=1.26.1
matplotlib>=3.8.1
sentence-transformers>=2.2.2
pi-heif>=0.13.0 